requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
async-exit-stack==1.0.1
async-generator==1.10
beautifulsoup4==4.9.1
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
dnspython==2.0.0
email-validator==1.1.1
extruct==0.10.0
fastapi==0.61.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
html-text==0.5.2
html5lib==1.1
httptools==0.1.1
idna==2.10
isodate==0.6.0
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
jstyleson==0.0.2
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mf2py==1.1.2
mongoengine==0.21.0
mypy-extensions==0.4.3
pathspec==0.8.0
promise==2.3
pydantic==1.6.1
pylint==2.6.0
pymongo==3.11.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
python-slugify==4.0.1
pytz==2020.4
PyYAML==5.3.1
rdflib==4.2.2
rdflib-jsonld==0.5.0
regex==2020.7.14
requests==2.24.0
Rx==1.6.1
scrape-schema-recipe==0.1.1
six==1.15.0
soupsieve==2.0.1
starlette==0.13.6
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==3.1.0
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
validators==0.18.0
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
wrapt==1.12.1
